requirements.txt

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
WTForms==3.1.2
email-validator==2.2.0
Flask-Login==0.6.3
cryptography==43.0.3
openpyxl==3.1.5
reportlab==4.2.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
WTForms==3.1.2
email-validator==2.2.0
Flask-Login==0.6.3
cryptography==43.0.3
openpyxl==3.1.5
reportlab==4.2.2
gunicorn==22.0.0
